Setup build procedure
