+++ bots/base_bot/bot_type.h
@@ -6,7 +6,7 @@
  * (C) David Norman 2002 david@ellought.demon.co.uk
  *
  * This software may be reused for non-commercial purposes without charge, and
- * without notifying the author. Use of any part of this software for commercial 
+ * without notifying the author. Use of any part of this software for commercial
  * purposes without permission from the Author is prohibited.
  *
  * Modified by John Newbury
@@ -16,19 +16,23 @@
 
 /*******************************************************************************************
  * Change the #include to specify the header for the main class of your Bot.               *
- * Change the typedef to make BOT_TYPE specify the main class of your Bot.                 *
+ * Change the using to make BOT_TYPE specify the main class of your Bot.                 *
  * Change the #define BOT_FAMILY to specify the family (generic name) of your bot.         *
  * Change the #define BOT_GENERATION to specify the generation (base version) of your Bot. *
  *******************************************************************************************/
 
-#ifndef DAIDE_CLIENT_BOTS_RANDBOT_BOT_TYPE_H
-#define DAIDE_CLIENT_BOTS_RANDBOT_BOT_TYPE_H
+#ifndef DAIDE_CLIENT_BOTS_BASEBOT_BOT_TYPE_H
+#define DAIDE_CLIENT_BOTS_BASEBOT_BOT_TYPE_H
 
-#include "randbot.h"
+#include "base_bot.h"
 
-using BOT_TYPE = RandBot;
+namespace DAIDE {
 
-#define BOT_FAMILY "RandBot"
+using BOT_TYPE = BaseBot;
+
+#define BOT_FAMILY "BaseBot"
 #define BOT_GENERATION "8~2~b"
 
-#endif // DAIDE_CLIENT_BOTS_RANDBOT_BOT_TYPE_H
+} // namespace DAIDE
+
+#endif // DAIDE_CLIENT_BOTS_BASEBOT_BOT_TYPE_H
