+++ bots/hold_bot/hold_bot.h
@@ -6,7 +6,7 @@
  * (C) David Norman 2002 david@ellought.demon.co.uk
  *
  * This software may be reused for non-commercial purposes without charge, and
- * without notifying the author. Use of any part of this software for commercial 
+ * without notifying the author. Use of any part of this software for commercial
  * purposes without permission from the Author is prohibited.
  *
  * Modified by John Newbury
@@ -17,7 +17,7 @@
 #ifndef DAIDE_CLIENT_BOTS_HOLDBOT_HOLDBOT_H
 #define DAIDE_CLIENT_BOTS_HOLDBOT_HOLDBOT_H
 
-#include "daide_client/base_bot.h"
+#include "base_bot.h"
 
 class HoldBot : public DAIDE::BaseBot {
     HoldBot(const HoldBot &other) = delete;                 // Copy constructor
