+++ bots/rand_bot/bot_type.h
@@ -6,7 +6,7 @@
  * (C) David Norman 2002 david@ellought.demon.co.uk
  *
  * This software may be reused for non-commercial purposes without charge, and
- * without notifying the author. Use of any part of this software for commercial 
+ * without notifying the author. Use of any part of this software for commercial
  * purposes without permission from the Author is prohibited.
  *
  * Modified by John Newbury
@@ -21,14 +21,18 @@
  * Change the #define BOT_GENERATION to specify the generation (base version) of your Bot. *
  *******************************************************************************************/
 
-#ifndef DAIDE_CLIENT_BOTS_HOLDBOT_BOT_TYPE_H
-#define DAIDE_CLIENT_BOTS_HOLDBOT_BOT_TYPE_H
+#ifndef DAIDE_CLIENT_BOTS_RANDBOT_BOT_TYPE_H
+#define DAIDE_CLIENT_BOTS_RANDBOT_BOT_TYPE_H
 
-#include "holdbot.h"
+#include "rand_bot.h"
 
-using BOT_TYPE = HoldBot;
+namespace DAIDE {
 
-#define BOT_FAMILY "HoldBot"
+using BOT_TYPE = RandBot;
+
+#define BOT_FAMILY "RandBot"
 #define BOT_GENERATION "8~2~b"
 
-#endif // DAIDE_CLIENT_BOTS_HOLDBOT_BOT_TYPE_H
+} // namespace DAIDE
+
+#endif // DAIDE_CLIENT_BOTS_RANDBOT_BOT_TYPE_H
