--- daide_client/ai_client.cpp
@@ -1,218 +1,0 @@
-//// AiClient.cpp : Defines the entry point for the application.
-
-//// Copyright (C) 2012, John Newbury. See "Conditions of Use" in johnnewbury.co.cc/diplomacy/conditions-of-use.htm.
-
-//// Release 8~2~b
-
-//#include <winsock2.h>
-//#include <ws2tcpip.h>
-#include <iostream>
-#include <sstream>
-#include "ai_client.h"
-//#include "error_log.h"
-//#include "socket.h"
-
-//#define MAX_LOADSTRING 100
-
-//// Need to link with Ws2_32.lib
-//#pragma comment(lib, "ws2_32.lib")
-
-//// Global Variables:
-//HINSTANCE hInst;                                // current instance
-//TCHAR szTitle[MAX_LOADSTRING];                  // The title bar text
-//TCHAR szWindowClass[MAX_LOADSTRING];            // the main window class name
-
-//// Forward declarations of functions included in this code module:
-//ATOM MyRegisterClass(HINSTANCE hInstance);
-//BOOL InitInstance(HINSTANCE, int);
-//LRESULT CALLBACK WndProc(HWND, UINT, WPARAM, LPARAM);
-
-using DAIDE::BOT_TYPE;
-using DAIDE::the_bot;
-
-BOT_TYPE DAIDE::the_bot;
-//HWND main_wnd;
-
-int main(int argc, char *argv[])
-{
-    std::stringstream sstr;
-    for (int i = 0, e = argc; i < e; ++i) {
-        sstr << argv[i] << " ";
-    }
-
-    if (!the_bot.initialize(sstr.str())) {
-        std::cerr << "Couldn't initialize Bot" << std::endl;
-        return 1;
-    }
-
-    // Main message loop:
-    while (the_bot.is_active()) {
-        the_bot.m_socket.ReceiveData();
-        the_bot.OnSocketMessage();
-    }
-
-    return 0;
-}
-
-//int APIENTRY _tWinMain(HINSTANCE hInstance,
-//                       HINSTANCE hPrevInstance,
-//                       LPTSTR lpCmdLine,
-//                       int nCmdShow
-//                       )
-//{
-//    UNREFERENCED_PARAMETER(hPrevInstance);
-//    UNREFERENCED_PARAMETER(lpCmdLine);
-
-//    _CrtSetDbgFlag(_CRTDBG_ALLOC_MEM_DF
-//                   | _CRTDBG_LEAK_CHECK_DF); // report memory leaks in Debug mode
-
-//    // TODO: Place code here.
-//    MSG msg;
-//    HACCEL hAccelTable;
-
-//    // Initialize global strings
-//    LoadString(hInstance, IDC_AICLIENT, szWindowClass, MAX_LOADSTRING);
-//    MyRegisterClass(hInstance);
-
-//    // Perform application initialization:
-//    if (!InitInstance(hInstance, nCmdShow))
-//    {
-//        return FALSE;
-//    }
-
-//    hAccelTable = LoadAccelerators(hInstance, MAKEINTRESOURCE(IDC_AICLIENT));
-
-//    if (!the_bot.initialize())
-//    {
-//        display("Couldn't initialize Bot");
-//        the_bot.end_dialog();
-//    }
-
-//    // Main message loop:
-//    while (GetMessage(&msg, NULL, 0, 0))
-//    {
-//        if (!TranslateAccelerator(msg.hwnd, hAccelTable, &msg))
-//        {
-//            TranslateMessage(&msg);
-//            DispatchMessage(&msg);
-//        }
-//    }
-
-//    WSACleanup();
-
-//    return (int) msg.wParam;
-//}
-
-////
-////  FUNCTION: MyRegisterClass()
-////
-////  PURPOSE: Registers the window class.
-////
-////  COMMENTS:
-////
-////    This function and its usage are only necessary if you want this code
-////    to be compatible with Win32 systems prior to the 'RegisterClassEx'
-////    function that was added to Windows 95. It is important to call this function
-////    so that the application will get 'well formed' small icons associated
-////    with it.
-////
-//ATOM MyRegisterClass(HINSTANCE hInstance)
-//{
-//    WNDCLASSEX wcex;
-//    wcex.cbSize = sizeof(WNDCLASSEX);
-//    wcex.style = CS_HREDRAW | CS_VREDRAW;
-//    wcex.lpfnWndProc = WndProc;
-//    wcex.cbClsExtra = 0;
-//    wcex.cbWndExtra = 0;
-//    wcex.hInstance = hInstance;
-//    wcex.hIcon = LoadIcon(hInstance, MAKEINTRESOURCE(IDI_AICLIENT));
-//    wcex.hCursor = LoadCursor(NULL, IDC_ARROW);
-//    wcex.hbrBackground = (HBRUSH)(COLOR_WINDOW + 1);
-//    wcex.lpszMenuName = MAKEINTRESOURCE(IDC_AICLIENT);
-//    wcex.lpszClassName = szWindowClass;
-//    wcex.hIconSm = LoadIcon(wcex.hInstance, MAKEINTRESOURCE(IDI_SMALL));
-
-//    return RegisterClassEx(&wcex);
-//}
-
-////
-////   FUNCTION: InitInstance(HINSTANCE, int)
-////
-////   PURPOSE: Saves instance handle and creates main window
-////
-////   COMMENTS:
-////
-////        In this function, we save the instance handle in a global variable and
-////        create and display the main program window.
-////
-
-//BOOL InitInstance(HINSTANCE hInstance,
-//                  int nCmdShow
-//                  )
-//{
-//    hInst = hInstance; // Store instance handle in our global variable
-
-//    main_wnd = CreateWindow(szWindowClass, szTitle, WS_OVERLAPPEDWINDOW,
-//                            CW_USEDEFAULT, 0, CW_USEDEFAULT, 0, NULL, NULL, hInstance, NULL);
-
-//    if (!main_wnd)
-//    {
-//        return FALSE;
-//    }
-
-//    SetWindowPos(main_wnd,                    // handle to window
-//                 0,                           // placement-order handle
-//                 0,                           // horizontal position
-//                 0,                           // vertical position
-//                 250,                         // width
-//                 60,                          // height
-//                 SWP_NOZORDER | SWP_NOMOVE    // window-positioning flags
-//                 );
-
-//    ShowWindow(main_wnd, nCmdShow);
-//    UpdateWindow(main_wnd);
-
-//    return TRUE;
-//}
-
-////
-////  FUNCTION: WndProc(HWND, UINT, WPARAM, LPARAM)
-////
-////  PURPOSE:  Processes messages for the main window.
-////
-////  WM_COMMAND	- process the application menu
-////  WM_PAINT	- Paint the main window
-////  WM_DESTROY	- post a quit message and return
-////
-////
-//LRESULT CALLBACK WndProc(HWND hWnd,
-//                         UINT message,
-//                         WPARAM wParam,
-//                         LPARAM lParam
-//                         )
-//{
-//    PAINTSTRUCT ps;
-//    HDC hdc;
-
-//    switch (message)
-//    {
-//    case WM_PAINT:
-//        hdc = BeginPaint(hWnd, &ps);
-//        // TODO: Add any drawing code here...
-//        EndPaint(hWnd, &ps);
-//        break;
-//    case WM_SOCKET_STATE:
-//        Socket::OnSocketState(wParam, lParam);
-//        break;
-//    case WM_SOCKET_MESSAGE:
-//        the_bot.OnSocketMessage();
-//        break;
-//    case WM_DESTROY:
-//        PostQuitMessage(0);
-//        break;
-//    default:
-//        return DefWindowProc(hWnd, message, wParam, lParam);
-//    }
-//    return 0;
-//}
-
